python/requirements.txt
fastapi 
uvicorn
googlesearch-python
beautifulsoup4
requests
